requirements.txt
cloudant
python-dotenv
flask
flask_httpauth
influxdb_client
jsonify
shortuuid